--- api/reviews.json
@@ -1,3 +1,0 @@
-{
-    "reviews": []
-}
